spycats/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
psycopg==3.2.3
psycopg-binary==3.2.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
